git-svn-id: https://svn.jboss.org/repos/labs/labs/jbossrules/branches/DroolsChance@26894 c60d74c8-e8f6-0310-9e8f-d4a2fc68ab70

+++ drools-core/src/main/java/org/drools/degrees/factory/IDegreeFactory.java
@@ -17,7 +17,7 @@ public interface IDegreeFactory {
 	public boolean isClosedWorldAssumption();
 	
 	
-	public IDegree buildDegree(float val);
+	public IDegree buildDegree(double val);
 	
 	
 	public IDegree True();
